Move mania's HitWindows to osu.Game
