+++ osu.Game/Rulesets/Objects/HitWindows.cs
@@ -4,7 +4,7 @@
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Scoring;
 
-namespace osu.Game.Rulesets.Mania.Judgements
+namespace osu.Game.Rulesets.Objects
 {
     public class HitWindows
     {
